now the processing of the mismatched and matched works seperately with good accuracy
I finally got it to work to recognises matched faces from folders with even 2 or more faces in them. It will take only the first 2 and ignore the rest. Also, If the MTCNN does not recognise either of the faces, it will skip that entire folder so neither get added to the embedding lists. Also for the mismatched folder, the process works. It goes through over 6000 faces i think and it has to know that none of them are the the same people. but i have nt got both programs to work at the same time. yet...